gcp_weather_pulse/transformer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
python-dateutil==2.9.0.post0
